[judge, sandbox] implement Serialize and Deserialize for data structures inside `judge` and `sandbox` to allow driver program to pass these data structures conveniently. The additional dependencies introduced are declared as optional dependencies and features on these two crates.

+++ sandbox/src/daemon.rs
@@ -170,7 +170,7 @@ fn daemon_main(context: &ProcessDaemonContext) -> Result<ProcessExitStatus> {
             WaitStatus::Signaled(_, Signal::SIGUSR1, _) =>
                 return Err(Error::from(ErrorKind::ChildStartupFailed)),
             WaitStatus::Signaled(_, sig, _) =>
-                return Ok(ProcessExitStatus::KilledBySignal(sig)),
+                return Ok(ProcessExitStatus::KilledBySignal(sig as i32)),
             _ => ()
         };
